refactor(api/shallow): move services to subdirectory
