+++ api/lib/shallow.dart
@@ -2,9 +2,9 @@ export 'src/shallow/authentication/data.dart';
 export 'src/shallow/authentication/module.dart';
 export 'src/shallow/collection/filtering.dart';
 export 'src/shallow/collection/module.dart';
-export 'src/shallow/course.dart';
 export 'src/shallow/entity.dart';
-export 'src/shallow/news.dart';
-export 'src/shallow/school.dart';
+export 'src/shallow/services/course.dart';
+export 'src/shallow/services/news.dart';
+export 'src/shallow/services/school.dart';
+export 'src/shallow/services/user.dart';
 export 'src/shallow/shallow.dart';
-export 'src/shallow/user.dart';
