+++ api/lib/src/shallow/services/course.dart
@@ -1,13 +1,13 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:time_machine/time_machine.dart';
 
-import 'collection/filtering.dart';
-import 'collection/module.dart';
-import 'entity.dart';
+import '../collection/filtering.dart';
+import '../collection/module.dart';
+import '../entity.dart';
+import '../shallow.dart';
+import '../utils.dart';
 import 'school.dart';
-import 'shallow.dart';
 import 'user.dart';
-import 'utils.dart';
 
 part 'course.freezed.dart';
