+++ api/lib/src/shallow/services/school.dart
@@ -1,11 +1,11 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:time_machine/time_machine.dart';
 
-import 'collection/filtering.dart';
-import 'collection/module.dart';
-import 'entity.dart';
-import 'shallow.dart';
-import 'utils.dart';
+import '../collection/filtering.dart';
+import '../collection/module.dart';
+import '../entity.dart';
+import '../shallow.dart';
+import '../utils.dart';
 
 part 'school.freezed.dart';
